feat: Add delay option for push notification in main.ts
This commit modifies the `main.ts` file to add a delay option for sending push notifications. The delay is specified in seconds and can be passed as a parameter in the request body. If no delay is provided, the notification will be sent immediately. This change enhances the functionality of the application by allowing for delayed push notifications.